refactor(battle): split SetupBoard into Core + store wrapper
SetupBoard was already fully presentational (no useTeamStore coupling)
but mixed the presentational body with the public wrapper. Extract the
~340-line SVG renderer to SetupBoardCore.tsx and keep SetupBoard.tsx as
a thin pass-through. No behavior change; public <SetupBoard /> API is
unchanged, so the three /simulator layout callers (desktop, tablet,
mobile) continue to work without modification. The seam lets future
store-bound fields be injected in the wrapper without touching Core.
